+++ .github/workflows/test.yaml
@@ -39,4 +39,8 @@ jobs:
 
       - name: Test
         run: |
-          bazel test --jobs=$(nproc) //...
+          bazel test //...
+
+      - name: Build (test)
+        run: |
+          bazel build  //...
